requirements.txt
attrs==21.2.0
cffi==1.15.0
click==8.0.3
cloudpickle==2.0.0
cycler==0.10.0
Cython==0.29.24
glfw==2.3.0
gym==0.21.0
imageio==2.9.0
importlib-resources==5.3.0
iniconfig==1.1.1
kiwisolver==1.3.2
lockfile==0.12.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
packaging==21.0
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
pyglet==1.5.21
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.7.1
six==1.16.0
toml==0.10.2
tqdm==4.62.3
zipp==3.6.0
